[ci] Fix isuses, add more custom build tests

+++ src/addons/journal.c
@@ -16,7 +16,7 @@ static char* flecs_journal_entitystr(
     if (_path && !strchr(_path, '.')) {
         path = flecs_asprintf("#[blue]%s", _path);
     } else {
-        uint32_t gen = entity >> 32;
+        uint32_t gen = (uint32_t)(entity >> 32);
         if (gen) {
             path = flecs_asprintf("#[normal]_%u_%u", (uint32_t)entity, gen);
         } else {
